% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakdown_by_rating.R
\name{breakdown_by_rating}
\alias{breakdown_by_rating}
\title{Returns ratings, entities, and households rated by PD by rating level in given timeframe.}
\usage{
breakdown_by_rating(from, to)
}
\arguments{
\item{from}{Start date, in yyyymmdd format}

\item{to}{End date, in yyyymmdd format}
}
\description{
Returns ratings, entities, and households rated by PD by rating level in given timeframe.
}
\examples{
breakdown_by_rating(from = 20170701, to = 20180630)
}
