% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_identified.R
\name{capacity_identified}
\alias{capacity_identified}
\title{Returns ratings, entities, households rated, and capacity identified (in dollars) by PD by fiscal year in given timeframe.}
\usage{
capacity_identified(from, to)
}
\arguments{
\item{from}{Start date, in yyyymmdd format}

\item{to}{End date, in yyyymmdd format}
}
\description{
Returns ratings, entities, households rated, and capacity identified (in dollars) by PD by fiscal year in given timeframe.
}
\examples{
capacity_identified(from = 20150701, to = 20180630)
}
