% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_year.R
\name{class_year}
\alias{class_year}
\title{Returns class year (by decade) of prospects rated by PD in given timeframe.}
\usage{
class_year(from, to)
}
\arguments{
\item{from}{Start date, in yyyymmdd format}

\item{to}{End date, in yyyymmdd format}
}
\description{
Returns class year (by decade) of prospects rated by PD in given timeframe.
}
\examples{
class_year(from = 20170701, to = 20180630)
}
