% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gift_purpose.R
\name{gift_purpose}
\alias{gift_purpose}
\title{Returns top n gift purposes for rated prospects and number of gifts made in given timeframe.}
\usage{
gift_purpose(from, to, n_purpose)
}
\arguments{
\item{from}{Start date, in yyyymmdd format}

\item{to}{End date, in yyyymmdd format}

\item{n_purpose}{Number of gift purposes to return}
}
\description{
Returns top n gift purposes for rated prospects and number of gifts made in given timeframe.
}
\examples{
gift_purpose(from = 20170701, to = 20180630, n_purpose = 5)
}
