% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/international_households.R
\name{international_households}
\alias{international_households}
\title{Returns number of international households in given timeframe.}
\usage{
international_households(from, to)
}
\arguments{
\item{from}{Start date, in yyyymmdd format}

\item{to}{End date, in yyyymmdd format}
}
\description{
Returns number of international households in given timeframe.
}
\examples{
international_households(from = 20150701, to = 20180630)
}
