% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_rating.R
\name{median_rating}
\alias{median_rating}
\title{Returns median and average rating by fiscal year in given timeframe.}
\usage{
median_rating(from, to)
}
\arguments{
\item{from}{Start date, in yyyymmdd format}

\item{to}{End date, in yyyymmdd format}
}
\description{
Returns median and average rating by fiscal year in given timeframe.
}
\examples{
median_rating(from = 20150701, to = 20180630)
}
