% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phil_affin_updates.R
\name{phil_affin_updates}
\alias{phil_affin_updates}
\title{Returns philanthropic affinities added by PD by fiscal year in given timeframe.}
\usage{
phil_affin_updates(from, to)
}
\arguments{
\item{from}{Start date, in yyyymmdd format}

\item{to}{End date, in yyyymmdd format}
}
\description{
Returns philanthropic affinities added by PD by fiscal year in given timeframe.
}
\examples{
phil_affin_updates(from = 20170701, to = 20180630)
}
