% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_supported.R
\name{proposals_created}
\alias{proposals_created}
\title{Returns development officers supported and number of proposals created for each DO in given timeframe.}
\usage{
proposals_created(from, to)
}
\arguments{
\item{from}{Start date, in yyyymmdd format}

\item{to}{End date, in yyyymmdd format}
}
\description{
Returns development officers supported and number of proposals created for each DO in given timeframe.
}
\examples{
proposals_created(from = 20170701, to = 20180630)
}
