% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospect_info.R
\name{prospect_info}
\alias{prospect_info}
\title{Returns rated households, percentage of those households contacted in past year,}
\usage{
prospect_info(from, to)
}
\arguments{
\item{from}{Start date, in yyyymmdd format}

\item{to}{End date, in yyyymmdd format}
}
\description{
Returns rated households, percentage of those households contacted in past year,
}
\examples{
prospect_info(from = 20170701, to = 20180630)
}
