% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_supported.R
\name{total_units_supported}
\alias{total_units_supported}
\title{Returns sum of development officers and units supported in given timeframe.}
\usage{
total_units_supported(from, to)
}
\arguments{
\item{from}{Start date, in yyyymmdd format}

\item{to}{End date, in yyyymmdd format}
}
\description{
Returns sum of development officers and units supported in given timeframe.
}
\examples{
total_units_supported(from = 20170701, to = 20180630)
}
