% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_supported.R
\name{units_supported}
\alias{units_supported}
\title{Returns units supported and number of development officers in each unit in given timeframe.}
\usage{
units_supported(from, to)
}
\arguments{
\item{from}{Start date, in yyyymmdd format}

\item{to}{End date, in yyyymmdd format}
}
\description{
Returns units supported and number of development officers in each unit in given timeframe.
}
\examples{
units_supported(from = 20170701, to = 20180630)
}
